Add function sortNub, nub with sorted output
In commit a86144bad6 the behavior of `ordNub` changed so that the
output after removal of duplicates is no longer sorted. The new
behavior is correct, but some software was relying on the old
behavior. Code that depends on the output being sorted can switch
to use this new function.